requirements.txt
django>=3.2.14
asgiref==3.7.2
attrs==23.1.0
django-cors-headers==4.1.0
django-filter==23.3
djangorestframework==3.14.0
drf-nested-routers==0.93.4
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.17.3
jsonschema-specifications==0.1.3
Markdown==3.4.4
mysqlclient==2.1.1
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.8.9
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
